Initial Commit
This is my first game program, where I've set out to recreate the
classic game Pong.  This initial commit shows the code to be very
sloppy.  However, I plan on re-organizing the code so that it's up to
par with today's standard practices.  I also wish to implement a
cpu-controlled player option.