Created Entity Class
With the help of a good friend of mine, I've created & implemented the
"Entity" class to contain the basic functions of the ball & paddle
entities.  The next step will be to create a Ball subclass & Paddle
subclass to further reduce the amount of redundancy that's currently in
Game1.cs, and thus make the program more efficient and much more
professionally laid out.